requirements.txt
playwright
apscheduler
python-telegram-bot
structlog
polars
python-dotenv
tenacity
notebooklm-mcp-cli
httpx
beautifulsoup4
pydantic-settings
feedparser
fastapi
uvicorn
